Unify UnconfirmedMessageHandler interface in pkg/retry/handler
The node and VM relay packages each defined near-identical UMH
interfaces (differing only by Stopped()). Move the canonical
definition to pkg/retry/handler, co-located with the concrete
implementation, so both consumers share a single contract. The
VM relay interface now includes Stopped() for parity with the
node path.
